prevent deletion of .svn folder

+++ src/main/java/de/shaoranlaos/scm_backup_plugin/BackupRunner.java
@@ -67,7 +67,9 @@ public class BackupRunner implements Runnable {
 		} else {
 			LOG.info("Delete old backupfiles.");
 			for (File delFile : localBackupDir.listFiles()) {
-				delFile.delete();
+				if (!delFile.getName().equals(".svn")) {
+					delFile.delete();
+				}
 			}
 		}
